tests/wpt/web-platform-tests/bluetooth/requestDevice/canonicalizeFilter/wrong-service-in-optionalServices-member.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-test.js"></script>
<script src="/bluetooth/resources/bluetooth-fake-devices.js"></script>
<script>
'use strict';
const test_desc = 'Invalid optional service must reject the promise.';
const expected = new TypeError();
const test_specs = [
  {optionalServices: ['wrong_service'], filters: [{services: ['heart_rate']}]},
  {
    optionalServices: ['wrong_service'],
    filters: [{services: ['heart_rate'], name: 'Name'}]
  },
  {
    optionalServices: ['wrong_service'],
    filters: [{services: ['heart_rate'], namePrefix: 'Pre'}]
  },
  {
    optionalServices: ['wrong_service'],
    filters: [{services: ['heart_rate'], name: 'Name', namePrefix: 'Pre'}]
  },
  {optionalServices: ['wrong_service'], filters: [{name: 'Name'}]}, {
    optionalServices: ['wrong_service'],
    filters: [{name: 'Name', namePrefix: 'Pre'}]
  },
  {optionalServices: ['wrong_service'], filters: [{namePrefix: 'Pre'}]}
];

bluetooth_test(() => {
  let test_promises = Promise.resolve();
  test_specs.forEach(args => {
    test_promises = test_promises.then(
        () => assert_promise_rejects_with_message(
            requestDeviceWithTrustedClick(args), expected));
  });
  return test_promises;
}, test_desc);
</script>
